fixed lookup dynamic variables
