+++ LookdownComponent/Lookdown/components/LookdownControlClassic.tsx
@@ -19,6 +19,7 @@ import { useEntityOptions, useLanguagePack, useMetadata } from "../services/Data
 import { getClassicDropdownOptions } from "../services/DropdownHelper";
 import { getCustomFilterString, getHandlebarsVariables } from "../services/TemplateService";
 import { EntityOption, LookdownControlProps, OpenRecordMode } from "../types/typings";
+import { useAttributeOnChange } from "../hooks/useAttributeOnChange";
 
 const DEFAULT_BORDER_STYLES: IStyle = {
   borderColor: "#666",
@@ -111,6 +112,8 @@ export default function LookdownControlClassic({
     showIcon,
     iconSize
   );
+
+  useAttributeOnChange(metadata, customFilter);
 
   const isError = isErrorLanguagePack || isErrorMetadata || isErrorEntityOptions;
 
@@ -272,19 +275,6 @@ export default function LookdownControlClassic({
       },
     ],
   };
-
-  const invalidateFetchDataFn = useCallback((context) => {
-    queryClient.invalidateQueries({ queryKey: ["entityRecords"] });
-  }, []);
-
-  useEffect(() => {
-    const templateVar = getHandlebarsVariables((metadata?.lookupView.fetchxml ?? "") + (customFilter ?? ""));
-    if (templateVar.length) {
-      templateVar.forEach((v) => {
-        Xrm.Page.data.entity.attributes.get(v)?.addOnChange(invalidateFetchDataFn);
-      });
-    }
-  }, [metadata?.lookupView.fetchxml, customFilter]);
 
   return (
     <Stack horizontal styles={{ root: { width: "100%" } }}>
